requirements.txt

djangorestframework
django-filter
